Add files via upload
Create a Triangle class. Add exception handling statements to ensure the following
conditions: all sides are greater than 0 and sum of any two sides is greater than the third
side. The class should also have overloaded functions for calculating the area of a right
angled triangle as well as using Heron's formula to calculate the area of any type of
triangle.