Rename h.cpp to C1.cpp
Write a program to compute the sum of the first n terms of the following series:

S = 1-1/2^2+1/3^3-....1/n^n

The number of terms n is to be taken from the user through the command line. If the
command line argument is not found then prompt the user to enter the value of n.